refactor: move ShipmentCostViewForm to vehicleInOutReport and update related imports
